requirement.txt
gspread>=3.7.0
google-auth>=1.35.0
gspread-dataframe>=3.1.0
pandas>=1.3.0
beautifulsoup4>=4.9.3
selenium>=4.0.0
requests>=2.25.1
